casts

+++ src/MCP3008.cpp
@@ -96,7 +96,10 @@ unsigned short MCP3008::read(const std::uint8_t channel) const {
         throw std::runtime_error("spi transfer failed");
     }
 
-    return ((rxData[1] << 8) | rxData[2]) & 0x3ff;
+    return (
+        (static_cast<unsigned short>(rxData[1]) << 8) |
+         static_cast<unsigned short>(rxData[2])
+    ) & 0x3ff;
 
 }
